Rename versioneer.py to _versioneer.py
That way it doesn't break tab completion for versioned_hdf5/.
